+++ examples/web/src/sse/controller.rs
@@ -1,6 +1,7 @@
-use std::time::Duration;
-
 use async_stream::stream;
+use std::time::Duration;
+use tokio::time::sleep;
+
 use sword::prelude::*;
 use sword::web::*;
 
@@ -9,15 +10,16 @@ pub struct SseController;
 
 impl SseController {
     #[sse("/countdown")]
-    async fn countdown(&self) -> SseResult {
+    async fn countdown(&self) -> Sse<impl EventStream + use<>> {
         let events = stream! {
             for i in (1..=5).rev() {
-                tokio::time::sleep(Duration::from_millis(250)).await;
+                sleep(Duration::from_millis(250)).await;
                 yield Ok(Event::default().event("countdown").data(i.to_string()));
             }
-            yield Ok(Event::default().event("done").data("lift off!"));
+
+            yield Ok(Event::default().event("done").data("The countdown has finished!"));
         };
 
-        Sse::new(Box::pin(events))
+        Sse::new(events)
     }
 }